// src/@types/deezer_api.d.ts
// Generated by https://quicktype.io

interface DeezerTrackResults {
  id: number;
  readable: boolean;
  title: string;
  title_short: string;
  title_version: string;
  link: string;
  duration: number;
  rank: number;
  explicit_lyrics: boolean;
  explicit_content_lyrics: number;
  explicit_content_cover: number;
  preview: string;
  md5_image: string;
  artist: DeezerArtist;
  contributors: DeezerArtist[];
  album: DeezerAlbum;
  type: string;
}

// ? Deezer track data api types

export interface DeezerTrackDataAPI extends DeezerTrackResults {
  isrc: string;
  share: string;
  track_position: number;
  disk_number: number;
  release_date: string;
  bpm: number;
  gain: number;
  available_countries: string[];
}

// ? Deezer track data results api types

export interface DeezerTrackResultsAPI {
  data: DeezerTrackResults[];
  total: number;
}

interface DeezerAlbum {
  id: number;
  title: string;
  link?: string;
  cover: string;
  cover_small: string;
  cover_medium: string;
  cover_big: string;
  cover_xl: string;
  md5_image: string;
  release_date?: string;
  tracklist: string;
  type: string;
}

interface DeezerArtist {
  id: number;
  name: string;
  link: string;
  share: string;
  picture: string;
  picture_small: string;
  picture_medium: string;
  picture_big: string;
  picture_xl: string;
  radio?: boolean;
  tracklist: string;
  type: string;
  role?: string;
}

// ? Deezer artist info api types

export interface DeezerArtistInfoApi {
  data: DeezerArtistInfo[];
  total: number;
  next: string;
}

interface LastFMArtistInfo {
  name: string;
  mbid: string;
  url: string;
  image: {
    text: string;
    size: 'small' | 'medium' | 'large' | 'extralarge' | 'mega';
  }[];
  streamable: '0' | '1';
  ontour: '0' | '1';
  stats: {
    listeners: number;
    playcount: number;
  };
  similar: LastFMArtistInfo[];
  tags: {
    tag: { name: string; url: string }[];
  };
  bio: {
    links: {
      link: {
        text: string;
        rel: string;
        url: string;
      };
    };
    published: string;
    summary: string;
    content: string;
  };
}

export interface DeezerArtistInfo extends OnlineArtistArtworks {
  id: number;
  name: string;
  link: string;
  picture: string;
  picture_big: string;
  picture_xl: string;
  nb_album: number;
  nb_fan: string;
  radio: boolean;
  tracklist: string;
  type: string;
}